+++ ebpub/ebpub/streets/models.py
@@ -520,6 +520,14 @@ class PlaceType(models.Model):
     map_icon_url = models.TextField(blank=True, null=True)
     map_color = models.CharField(max_length=255, blank=True, null=True, help_text="CSS Color used on maps to display this type of place. eg #FF0000")
 
+    def get_map_icon_url(self):
+        from django.conf import settings
+        url = self.map_icon_url or u''
+        if url and not (url.startswith('/') or url.startswith('http')):
+            url = '%s/%s' % (settings.STATIC_URL.rstrip('/'), url)
+        return url
+
+
     def natural_key(self):
         return (self.slug, )